+++ launch/patsecure-launcher.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-exec pkexec deepin-terminal -e "bash -c '/home/Pattoo/PatSecure/patsecure.sh'"
-